Final Commit

+++ Backend/src/main/java/de/cerberus/backend/repository/PcRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import de.cerberus.backend.entity.Pc;
 
-public interface PcRepository extends JpaRepository<Pc, Long> {
+public interface PcRepository extends JpaRepository<Pc, Integer> {
     List<Pc> findByPcPoolPcPoolId(Integer poolId);
 }
